lambda/requirements.txt
pillow
boto3

